rename and card changes
