+++ src/components/TitleBlock.tsx
@@ -1,30 +1,20 @@
 import styled from 'styled-components';
 
-interface TitleBlockProps {
-  content: {
-    title: string;
-    subtitle: string;
-    image: {
-      picture: any;
-      alt: string;
-    };
-    buttons: {
-      content: string;
-      link: string;
-    }[];
-  };
-}
+import { HomeBlockProps } from '../interfaces/HomeBlockProps';
 
-const TitleBlock = (props: TitleBlockProps) => (
+
+const TitleBlock = (props: HomeBlockProps) => (
   <TitleBlockDiv>
     <Title>{props.content.title}</Title>
     <Description>{props.content.subtitle}</Description>
     <Image src={props.content.image.picture} alt={props.content.image.alt} />
-    {props.content.buttons.map((button) => (
-      <a key={`t-${button.content}`} href={button.link}>
-        <Button>{button.content}</Button>
-      </a>
-    ))}
+    <div>
+      {props.content.buttons.map((button) => (
+        <a key={`t-${button.content}`} href={button.link}>
+          <Button>{button.content}</Button>
+        </a>
+      ))}
+    </div>
   </TitleBlockDiv>
 );
 
@@ -39,8 +29,8 @@ const TitleBlockDiv = styled.div`
     'Image'
     'Button';
 
-  @media ${(props) => `${props.theme.viewport.mobile}`} {
-    grid-template-columns: 3fr 2fr;
+  @media ${(props) => `${props.theme.viewport.tablet}`} {
+    grid-template-columns: 1fr 1fr;
     grid-template-areas:
       'Title Image'
       'Description Image'
@@ -50,6 +40,7 @@ const TitleBlockDiv = styled.div`
 
 const Title = styled.h1`
   grid-area: Title;
+  font-size: ${(props) => props.theme.fontSizes.large};
   font-family: ${(props) => props.theme.fonts.heading};
   margin: 0;
 `;
@@ -57,7 +48,8 @@ const Title = styled.h1`
 const Description = styled.p`
   grid-area: Description;
   margin: 0;
-`;
+  font-size: ${(props) => props.theme.fontSizes.regular};
+  `;
 
 const Image = styled.img`
   grid-area: Image;
@@ -67,11 +59,19 @@ const Image = styled.img`
 
 const Button = styled.button`
   grid-area: Button;
-  background: rgba(147, 201, 254, 1);
   border: 3px solid #0078e8;
-  border-radius: 5px;
   padding: 10px 30px;
+  background: rgba(255,255,255,0);
+  color: #0078e8;
+  margin-right: 20px;
   font-family: ${(props) => props.theme.fonts.content};
+  &:hover{
+    background: #0078e8;
+    color: white;
+    cursor:pointer;
+    transition: all .6s ease;
+    -webkit-transition: all .6s ease;
+  }
 `;
 
 export default TitleBlock;
